requirements-ui.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
